refactor clean
