+++ app/src/main/java/com/example/nikulin/presentation/adapters/LatestMemesAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.nikulin.ui.adapters
+package com.example.nikulin.presentation.adapters
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
