+++ app/src/main/java/com/example/nikulin/presentation/di/UseCaseModule.kt
@@ -1,0 +1,17 @@
+package com.example.nikulin.presentation.di
+
+import com.example.nikulin.domain.RemoteRepository
+import com.example.nikulin.domain.usecase.GetMemesUseCase
+import dagger.Module
+import dagger.Provides
+
+
+@Module
+object UseCaseModule {
+
+    @Provides
+    fun provideGetMemesUseCase(remoteRepository: RemoteRepository): GetMemesUseCase {
+        return GetMemesUseCase(remoteRepository)
+    }
+
+}
