+++ app/src/main/res/layout/fragment_hot_memes.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragments.HotMemesFragment">
+    tools:context=".presentation.fragments.HotMemesFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvHotMemes"
