+++ app/src/main/res/layout/fragment_latest_memes.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragments.LatestMemesFragment">
+    tools:context=".presentation.fragments.LatestMemesFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvLatestMemes"
