+++ app/src/main/res/layout/fragment_top_memes.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragments.TopMemesFragment">
+    tools:context=".presentation.fragments.TopMemesFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvBestMemes"
